Update external-services to inherit from superclass; Update imports
