+++ app/models/external-storage-service.ts
@@ -1,37 +1,8 @@
 import { attr } from '@ember-data/model';
 
-import OsfModel from './osf-model';
+import ExternalServiceModel from './external-service';
 
-export enum CredentialsFormat {
-    OAUTH = 'oauth',
-    OAUTH2 = 'oauth2',
-    URL_USERNAME_PASSWORD = 'url_username_password',
-    USERNAME_PASSWORD = 'username_password',
-    REPO_TOKEN = 'repo_token',
-    ACCESS_SECRET_KEYS = 'access_secret_keys',
-}
-
-export enum TermsOfServiceCapabilities {
-    ADD_UPDATE_FILES = 'add_update_files',
-    ADD_UPDATE_FILES_PARTIAL = 'add_update_files_partial',
-    DELETE_FILES = 'delete_files',
-    DELETE_FILES_PARTIAL = 'delete_files_partial',
-    FORKING = 'forking',
-    FORKING_PARTIAL = 'forking_partial',
-    LOGS = 'logs',
-    LOGS_PARTIAL = 'logs_partial',
-    PERMISSIONS = 'permissions',
-    PERMISSIONS_PARTIAL = 'permissions_partial',
-    REGISTERING = 'registering',
-    REGISTERING_PARTIAL = 'registering_partial',
-    FILE_VERSIONS = 'file_versions',
-}
-
-export default class ExternalStorageServiceModel extends OsfModel {
-    @attr('fixstring') name!: string;
-    @attr('string') credentialsFormat!: CredentialsFormat;
-    @attr('string') iconUrl!: string;
-    @attr('array') termsOfService!: TermsOfServiceCapabilities[];
+export default class ExternalStorageServiceModel extends ExternalServiceModel {
     @attr('boolean') configurableApiRoot!: boolean;
     @attr('number') maxConcurrentDownloads!: number;
     @attr('number') maxUploadMb!: number;
